requirements.txt
# Core dependencies
fastapi>=0.68.0
loguru>=0.5.3
pandas>=1.3.0
playwright>=1.22.0
pydantic>=1.8.2
pydantic-settings>=2.0.0
python-dotenv>=0.19.0
requests>=2.25.1
streamlit>=1.0.0
streamlit-extras>=0.2.0
python-multipart>=0.0.5
PyYAML>=6.0

# Web scraping and parsing
beautifulsoup4>=4.9.3
lxml>=4.6.3
parsel>=1.6.0
jmespath>=1.0.0
nested_lookup>=0.2.23

# Reddit API
praw>=7.5.0

# Text analysis
vaderSentiment>=3.3.2

# Data visualization
matplotlib>=3.4.2
plotly>=5.0.0
seaborn>=0.11.1
wordcloud>=1.8.1

# Database
supabase>=2.4.0
pymongo>=4.0.0

# Google AI
google-generativeai>=0.3.0

# Type hints
typing_extensions>=4.0.0

# Testing
pytest>=6.2.4
pytest-asyncio>=0.15.1